+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0] - Unreleased
+
+### Added
+
+- Added tab_size to Console and Text
+- Added protocol.is_renderable for runtime check
+
+### Changed
+
+- Console.markup attribute now effects Table
+- SeparatedConsoleRenderable and RichCast types
+
+### Fixed
+
+- Fixed tabs breaking rendering by converting to spaces
+
 ## [0.5.0] - 2020-02-23
 
 ### Changed